pyproject.toml
[project]
name = "hackaton-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain_gigachat",
    "langchain_openai",
    "langchain_core",
    "python-dotenv",
    "langgraph",
    "langchain_tavily",
    "pydantic",
    "langgraph_supervisor",
    "rich",
    "pyyaml",
    "loguru",
    "langchain_tavily",
    "langchain_ollama",
    "langchain_mcp_adapters",
    "langmem",
    "numpy",
    "ipykernel",
    



]

[tool.ruff]
# Set the maximum line length to 79.
line-length = 79

[tool.ruff.lint]
# Add the `line-too-long` rule to the enforced rule set. By default, Ruff omits rules that
# overlap with the use of a formatter, like Black, but we can override this behavior by
# explicitly adding the rule.
extend-select = ["E501"]